requirements.txt
Flask>=3.0.0
flask-cors>=5.0.0
Flask-SQLAlchemy>=3.1.0
Flask-JWT-Extended>=4.6.0
mediapipe>=0.10.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
groq>=0.9.0
Werkzeug>=3.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
gunicorn>=21.2.0
